refactored server code and removed portal fcb
