+++ package.json
@@ -58,16 +58,6 @@
     "onLanguage:html"
   ],
   "contributes": {
-    "configuration": {
-      "title": "powerplatform-vscode",
-      "properties": {
-        "powerplatform-vscode.enablePortalFeatures": {
-          "type": "boolean",
-          "default": false,
-          "description": "Enable Portal features in VSCode Extension"
-        }
-      }
-    },
     "commands": [
       {
         "command": "pacCLI.openDocumentation",
