+++ src/server/yamlServer.ts
@@ -16,20 +16,10 @@ import {
     WorkspaceFolder
 } from 'vscode-languageserver/node';
 import { URL } from 'url';
-import * as path from 'path';
-import * as fs from 'fs';
-import * as YAML from 'yaml';
-
+import { getEditedLineContent, getMatchedManifestRecords, IManifestElement } from './lib/portalAutoCompleteHelper';
 import {
     TextDocument
 } from 'vscode-languageserver-textdocument';
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const lineByLine = require('n-readlines');
-interface ManifestElement {
-    DisplayName: string;
-    RecordId: string;
-}
 
 // Create a connection for the server, using Node's IPC as a transport.
 // Also include all preview / proposed LSP features.
@@ -42,8 +32,6 @@ let hasConfigurationCapability = false;
 let hasWorkspaceFolderCapability = false;
 let hasDiagnosticRelatedInformationCapability = false;
 let workspaceRootFolder: WorkspaceFolder[] | null = null;
-const portalConfigFolderName = '.portalconfig';
-const manifest = '-manifest';
 
 
 connection.onInitialize((params: InitializeParams) => {
@@ -113,78 +101,19 @@ function getSuggestions(rowIndex: number, fileUrl: URL) {
     const completionItems: CompletionItem[] = [];
     if (matches) {
         const keyForCompletion = getKeyForCompletion(matches);
-        let matchedManifestRecords: ManifestElement[] = [];
-
-        const portalConfigFolderUrl = getPortalConfigFolderUrl() as URL | null; //https://github.com/Microsoft/TypeScript/issues/11498
-        if (portalConfigFolderUrl) {
-            const portalConfigFolderPath = portalConfigFolderUrl.href;
-            if (portalConfigFolderUrl && portalConfigFolderPath) {
-                const configFiles: string[] = fs.readdirSync(portalConfigFolderUrl);
-                configFiles.forEach(configFile => {
-                    if (configFile.includes(manifest)) { // this is based on the assumption that there will be only one manifest file in portalconfig folder
-                        const manifestFilePath = path.join(portalConfigFolderPath, configFile);
-                        const manifestData = fs.readFileSync(new URL(manifestFilePath), 'utf8');
-                        try {
-                            const parsedManifestData = YAML.parse(manifestData);
-                            matchedManifestRecords = parsedManifestData[keyForCompletion];
-                        } catch (exception) {
-                            // parsing failed. Add telemetry log.
-                        }
-                    }
-                })
-            }
-
-            if (matchedManifestRecords) {
-                matchedManifestRecords.forEach((element: ManifestElement) => {
-                    const item: CompletionItem = {
-                        label: element.DisplayName + " (" + element.RecordId + ")",
-                        insertText: element.RecordId,
-                        kind: CompletionItemKind.Value
-                    }
-                    completionItems.push(item);
-                });
-            }
+        const matchedManifestRecords: IManifestElement[] = getMatchedManifestRecords(workspaceRootFolder, keyForCompletion);
+        if (matchedManifestRecords) {
+            matchedManifestRecords.forEach((element: IManifestElement) => {
+                const item: CompletionItem = {
+                    label: element.DisplayName + " (" + element.RecordId + ")",
+                    insertText: element.RecordId,
+                    kind: CompletionItemKind.Value
+                }
+                completionItems.push(item);
+            });
         }
-
     }
     return completionItems;
-}
-
-function getPortalConfigFolderUrl() {
-    const workspaceRootFolderUri = workspaceRootFolder && workspaceRootFolder[0].uri;
-    let portalConfigFolderUrl = null;
-    if (workspaceRootFolderUri !== null) {
-        const workspaceRootFolderUrl = new URL(workspaceRootFolderUri);
-        const workspaceRootFolderContents: string[] = fs.readdirSync(workspaceRootFolderUrl);
-        for (let i = 0; i < workspaceRootFolderContents.length; i++) {
-            const fileName = workspaceRootFolderContents[i];
-            const filePath = path.join(workspaceRootFolderUrl.href, fileName);
-            const fileUrl = new URL(filePath);
-            const isDirectory = fs.statSync(fileUrl).isDirectory();
-            if (isDirectory && fileName === portalConfigFolderName) {
-                portalConfigFolderUrl = fileUrl;
-                return portalConfigFolderUrl;
-            }
-        }
-    }
-    return portalConfigFolderUrl;
-}
-
-function getEditedLineContent(rowIndex: number, fileUrl: URL) {
-    const liner = new lineByLine(fileUrl);
-    let line = liner.next();
-    let lineNumber = 0;
-    let userEditedLine = '';
-
-    while (line) {
-        if (lineNumber == rowIndex) {
-            userEditedLine = line.toString('ascii');
-            break;
-        }
-        line = liner.next();
-        lineNumber++;
-    }
-    return userEditedLine;
 }
 
 function getKeyForCompletion(matches: RegExpMatchArray) {
